requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
jinja2==3.1.4

google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

httpx==0.27.2
pytest==8.3.3
google-auth==2.36.0


python-dotenv==1.0.1
email-validator==2.3.0
requests==2.32.3

black==24.8.0
isort==5.13.2
flake8==7.1.1
pytest==8.3.3
